requirements.txt
fastapi
pydantic
python-dotenv
torch>=2.0.0
torchvision>=0.15.0
sentence-transformers
uvicorn[standard]
openai>=1.0.0
httpx>=0.27.0
pytest>=8.0.0
rembg>=2.0.0